Worked on integration. When you log in/register, UserManager creates a new User (with default customization choices) and sets this as the current user.
Created an instance of Game in each GameManager. Game stores the statistics of the game and an instance of Customization, which stores the customization choices of the current user.

Created a method called finishGame in AppManager. A mini game will call this method once it is finished, passing its Game instance as a paramter.